refactor: move `@plaiceholder/core` and `getImage` to plaiceholder
