+++ examples/next/src/pages/base64/multiple.tsx
@@ -1,8 +1,7 @@
 import * as React from "react";
 import { InferGetStaticPropsType } from "next";
 import Image from "next/image";
-import { getImage } from "@plaiceholder/next";
-import { getBase64 } from "plaiceholder";
+import { getBase64, getImage } from "plaiceholder";
 import { cx } from "@/styles";
 import { config } from "@/config";
 import { getAllPublicImagePaths } from "@/lib/images";
