+++ examples/next/src/pages/css/multiple.tsx
@@ -1,8 +1,7 @@
 import * as React from "react";
 import { InferGetStaticPropsType } from "next";
 import Image from "next/image";
-import { getImage } from "@plaiceholder/next";
-import { getCSS } from "plaiceholder";
+import { getCSS, getImage } from "plaiceholder";
 import { config } from "@/config";
 import { getAllPublicImagePaths } from "@/lib/images";
 import { cx } from "@/styles";
