--- packages/core/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "@plaiceholder/core",
-  "version": "1.0.0",
-  "description": "You probably shouldn't be using this directly.",
-  "author": "Joe Bell <joe@joebell.co.uk>",
-  "license": "Apache-2.0",
-  "funding": "https://plaiceholder.co",
-  "source": "src/index.ts",
-  "main": "dist/index.js",
-  "module": "dist/index.esm.js",
-  "typings": "dist/index.d.ts",
-  "files": [
-    "dist",
-    "README.md"
-  ],
-  "scripts": {
-    "build": "microbundle --external sharp --no-compress --target node --tsconfig tsconfig.json",
-    "dev": "yarn build -- watch",
-    "clean": "rimraf {dist,.rts2*}"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/joe-bell/plaiceholder.git"
-  },
-  "keywords": [
-    "lqip",
-    "placeholder",
-    "plaiceholder"
-  ],
-  "bugs": {
-    "url": "https://github.com/joe-bell/plaiceholder/issues"
-  },
-  "homepage": "https://github.com/joe-bell/plaiceholder#readme",
-  "peerDependencies": {
-    "sharp": ">= 0.26.2"
-  },
-  "dependencies": {
-    "image-size": "0.9.7"
-  }
-}
