+++ packages/next/package.json
@@ -32,13 +32,11 @@
     "url": "https://github.com/joe-bell/plaiceholder/issues"
   },
   "homepage": "https://github.com/joe-bell/plaiceholder#readme",
-  "dependencies": {
-    "@plaiceholder/core": "^1.0.0"
-  },
   "devDependencies": {
     "next": "10.2.2"
   },
   "peerDependencies": {
-    "next": ">= 10.0.0"
+    "next": ">= 10.0.0",
+    "sharp": ">= 0.26.2"
   }
 }
