+++ packages/plaiceholder/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./base64";
 export * from "./blurhash";
 export * from "./css";
+export * from "./get";
 export * from "./svg";
