+++ packages/tailwindcss/src/plugin.ts
@@ -1,13 +1,11 @@
 import plugin from "tailwindcss/plugin";
 import makeSynchronous from "make-synchronous";
 import cssEscape from "css.escape";
-import type { GetImage } from "@plaiceholder/core";
-import type { IGetCSS } from "plaiceholder";
+import type { GetImage, IGetCSS } from "plaiceholder";
 import { classNamePrefix } from "./config";
 
 const getPlaiceholder = makeSynchronous(async (imageUrl) => {
-  const { getImage } = require("@plaiceholder/next");
-  const { getCSS } = require("plaiceholder");
+  const { getCSS, getImage } = require("plaiceholder");
 
   const { buffer } = await (getImage as GetImage)(imageUrl);
   const pixelsCSS = await (getCSS as IGetCSS)(buffer);
